Move `useCalculateGriefingCollateral` to extra file
